refactor(types): разбор объявленных в описании типов — отдельный компонент
Разрешение типов параметров и возвращаемого значения из документирующего
комментария вынесено в DescribedTypeResolver: ConfigurationModuleMembersProvider
занимается регистрацией типов модулей, а не разбором описаний.
